### 2023-01-22 #5090 @jbytecode
👋👋👋 Dear @rusu24edward  👋👋👋

JOSS publishes articles about open source research software. The submission I'd like you to review is titled: "**Melodie: Agent-based Modeling in Python**". You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/5090).

Thank you in advance!
