### 2023-01-25 #5100 @editorialbot
[REVIEW]: Melodie: Agent-based Modeling in Python
**Submitting author:** @SongminYu (<a href="http://orcid.org/0000-0001-6062-4382">Songmin Yu</a>)
**Repository:** https://github.com/ABM4ALL/Melodie
**Branch with paper.md** (empty if default branch): JOSS-paper
**Version:** 0.7.0
