### 2023-01-26 677f206
Modified visualizer options and code
