### 2023-01-28 7a06939
Modified visualizer options and code
