### 2023-01-31 64c0026
Added dynamic action params getter
