+++ Melodie/boost/grid.pyx
@@ -9,6 +9,7 @@ cdef extern from "Python.h":
 import functools
 from typing import Type, Set, Dict, List, Tuple
 from Melodie.boost.vectorize import vectorize_2d
+
 from cpython.ref cimport PyObject
 from cpython cimport PyObject_GetAttr, PyObject_GetAttrString, \
     PyObject_GetItem, PyList_GetItem, PyList_Size, PyObject_SetAttr
@@ -24,6 +25,7 @@ from cython.operator cimport dereference as deref, preincrement as inc
 
 from .basics cimport Element, Agent
 from .agent_list cimport AgentList
+from .fastrand cimport randint
 cimport numpy as np
 import random
 import numpy as np
@@ -105,90 +107,6 @@ cdef class Spot(GridItem):
             "backgroundColor": "#ffffff"
         }
 
-
-cdef class AgentIDManager:
-    def __init__(self, long width, long height, bint allow_multi=False):
-        self._agents = [set() for i in range(width*height)]
-        self._width = width
-        self._height = height
-        self._max_id = 1000000
-        self.allow_multi = allow_multi
-        self._empty_spots = set()
-        self.all_categories = set()
-        for x in range(self._width):
-            for y in range(self._height):
-                self._empty_spots.insert(self._convert_to_1d(x, y))
-
-    cpdef add_agent(self, long agent_id, long category, long x, long y) except *:
-        self._add_agent(agent_id, category, x, y)
-
-    cdef void _add_agent(self, long agent_id, long category, long x, long y) except *:
-        cdef long pos_1d = self._convert_to_1d(x, y)
-        cdef cpp_set[long]* agents_on_spot = &self._agents[pos_1d]
-        cdef long agent_num_repr = self.agent_id_and_category_to_number(agent_id, category)
-        if (not self.allow_multi) and (agents_on_spot.count(pos_1d)!=0):  # If not allow multi, the spot should be empty before adding an agent.
-            raise ValueError(f"Multiple agents on one spot is not allowed.")
-        if agents_on_spot.count(agent_num_repr)!=0:
-            raise ValueError(f"Agent id {agent_id}, category {category} already on spot <{x}, {y}>")
-        
-        self._empty_spots.erase(pos_1d)
-        agents_on_spot.insert(agent_num_repr)
-        self.all_categories.insert(category)
-    
-    cpdef remove_agent(self, long agent_id, long category, long x, long y) except *:
-        self._remove_agent(agent_id, category, x, y)
-
-    cdef void _remove_agent(self, long agent_id, long category, long x, long y) except *:
-        cdef long pos_1d = self._convert_to_1d(x, y)
-        cdef cpp_set[long]* agents_on_spot = &self._agents[pos_1d]
-        cdef long agent_num_repr = self.agent_id_and_category_to_number(agent_id, category)
-        if agents_on_spot.count(agent_num_repr)==0:
-            raise ValueError(f"Agent id {agent_id}, category {category} does not exist on spot <{x}, {y}>")
-        agents_on_spot.erase(agent_num_repr)
-        if agents_on_spot.size()==0:
-            self._empty_spots.insert(pos_1d)
-
-    cdef long _convert_to_1d(self, long x, long y):
-        return x * self._height + y
-
-    cdef cpp_set[long]* _get_empty_spots(self):
-        return &self._empty_spots
-
-    cpdef set get_empty_spots(self):
-        return self._empty_spots
-
-    cpdef long agent_id_and_category_to_number(self, long agent_id, long category) except *:
-        if not 0<=agent_id<=self._max_id:
-            raise ValueError(f"Agent id {agent_id} out of range!")
-        return category*self._max_id + agent_id
-
-    @cython.cdivision(True)
[…]
