### 2023-02-01 ABM4ALL/Melodie#3 @rusu24edward
run_simulation bug
Reviewing the [Joss Submission](https://github.com/openjournals/joss-reviews/issues/5100), I am attempting to run the tutorials. I installed Melodie as per the installation instructions, and then ran `python3 run_simulator.py`. I get the following error in all of the tutorials:

```
Traceback (most recent call last):
  File "run_simulator.py", line 14, in <module>
    simulator.run()
  File "/Users/rusu1/tmp/Melodie/v_melodie/lib/python3.7/site-packages/Melodie/simulator.py", line 236, in run
    self.pre_run()
  File "/Users/rusu1/tmp/Melodie/v_melodie/lib/python3.7/site-packages/Melodie/simulator.py", line 99, in pre_run
    create_db_conn(self.config).clear_database()
  File "/Users/rusu1/tmp/Melodie/v_melodie/lib/python3.7/site-packages/MelodieInfra/db/db.py", line 122, in clear_database
    logger.info(f"Database contains tables: {self.connection.table_names()}.")
AttributeError: 'Engine' object has no attribute 'table_names'
```

I also get this error from `run_calibrator.py` in the CovidContagionCalibrator tutorial.
I also get this error from `run_trainer.py` in the RockPaperScissorsTrainer tutorial.
