### 2023-02-01 ABM4ALL/Melodie#4 @rusu24edward
run_analyzer bug
Similar to #3

`python3 run_analyzer.py` returns the following error:

```
Traceback (most recent call last):
  File "run_analyzer.py", line 7, in <module>
    analyzer.plot_health_state_share(id_scenario=0)
  File "/Users/rusu1/tmp/CovidContagion/source/analyzer.py", line 24, in plot_health_state_share
    df = self.db.read_dataframe("environment_result")
  File "/Users/rusu1/tmp/Melodie/v_melodie/lib/python3.7/site-packages/MelodieInfra/db/db.py", line 193, in read_dataframe
    return pd.read_sql(sql, self.connection)
  File "/Users/rusu1/tmp/Melodie/v_melodie/lib/python3.7/site-packages/pandas/io/sql.py", line 599, in read_sql
    pandas_sql = pandasSQL_builder(con)
  File "/Users/rusu1/tmp/Melodie/v_melodie/lib/python3.7/site-packages/pandas/io/sql.py", line 788, in pandasSQL_builder
    return SQLDatabase(con, schema=schema, meta=meta)
  File "/Users/rusu1/tmp/Melodie/v_melodie/lib/python3.7/site-packages/pandas/io/sql.py", line 1410, in __init__
    meta = MetaData(self.connectable, schema=schema)
TypeError: __init__() got multiple values for argument 'schema'
```
