### 2023-02-01 ABM4ALL/Melodie#3 @rusu24edward
I also get this error with ABMFrameworkComparison run_simulator
