### 2023-02-01 ABM4ALL/Melodie#5 @rusu24edward
Cannot launch MelodieStudio
After installing Melodie Studio with `pip3 install MelodieStudio` as per the [Visual Tutorial](https://abm4all.github.io/Melodie/html/gallery/covid_contagion_visual.html), I get the following error when I try to launch Melodie Studio with `python3 -m MelodieStudio`:
```
Traceback (most recent call last):
  File "/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/runpy.py", line 183, in _run_module_as_main
    mod_name, mod_spec, code = _get_module_details(mod_name, _Error)
  File "/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/runpy.py", line 142, in _get_module_details
    return _get_module_details(pkg_main_name, error)
  File "/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/runpy.py", line 109, in _get_module_details
    __import__(pkg_name)
  File "/Users/rusu1/tmp/Melodie/v_melodie/lib/python3.7/site-packages/MelodieStudio/__init__.py", line 7, in <module>
    from .main import studio_main
  File "/Users/rusu1/tmp/Melodie/v_melodie/lib/python3.7/site-packages/MelodieStudio/main.py", line 14, in <module>
    from .routes import (
  File "/Users/rusu1/tmp/Melodie/v_melodie/lib/python3.7/site-packages/MelodieStudio/routes/__init__.py", line 1, in <module>
    from .db_browser import db_browser
  File "/Users/rusu1/tmp/Melodie/v_melodie/lib/python3.7/site-packages/MelodieStudio/routes/db_browser.py", line 12, in <module>
    from MelodieStudio.routes.wrapper import wrapper_for_get
  File "/Users/rusu1/tmp/Melodie/v_melodie/lib/python3.7/site-packages/MelodieStudio/routes/wrapper.py", line 8, in <module>
    from MelodieStudio.models.messages import Response
  File "/Users/rusu1/tmp/Melodie/v_melodie/lib/python3.7/site-packages/MelodieStudio/models/__init__.py", line 2, in <module>
    from .pty import *
  File "/Users/rusu1/tmp/Melodie/v_melodie/lib/python3.7/site-packages/MelodieStudio/models/pty.py", line 1, in <module>
    from typing import TypedDict
ImportError: cannot import name 'TypedDict' from 'typing' (/Library/Frameworks/Python.framework/Versions/3.7/lib/python3.7/typing.py)
```
