### 2023-02-02 ABM4ALL/Melodie#5 @hzyrc6011
Thanks for reporting this bug. There is an error in our docs that Melodie is suitable for python>=3.8, not python>=3.7, so it is better to use python 3.8 or newer to test MelodieStudio.
