### 2023-02-02 ABM4ALL/Melodie#4 @hzyrc6011
There is an error in our docs that Melodie is suitable for python>=3.8, not python>=3.7, so it is better to use python 3.8 or newer to test this program.

Besides, this problem is very likely caused by error in  the version for sqlalchemy. Would you mind using this command:

```sh
python -m pip install sqlalchemy==1.4.41
```
to install the older sqlalchemy version and re-run the program?
