### 2023-02-02 #5100 @SongminYu
Hi @rusu24edward, thanks for your comments! My co-author replied to you about the issues in the Melodie repo. We made a mistake in the documentation. The requirement should be >= Python 3.8, not 3.7. Could you please try again with Python 3.8 or later? I hope the replies can solve the problems. Please let us know if it works. Thanks a lot!
