### 2023-02-08 dae279f
Modified melodie visualizer dependencies
