+++ Melodie/visualizer/visualizer.py
@@ -4,13 +4,10 @@ import asyncio
 import os
 import queue
 import shutil
-import sys
-from copy import deepcopy
 import threading
 import time
 from enum import Enum
 
-import websockets
 import json
 from queue import Queue
 from typing import Dict, Tuple, List, Any, Callable, Union, Set, TYPE_CHECKING, Optional
@@ -18,6 +15,7 @@ from typing import Dict, Tuple, List, Any, Callable, Union, Set, TYPE_CHECKING, 
 from websockets.exceptions import ConnectionClosedOK
 
 from .actions import Action
+from .visualizer_server import create_visualizer_server
 from .ws_protocol import MelodieVisualizerProtocol
 
 from MelodieInfra import OSTroubleShooter, get_sqlite_filename, MelodieExceptions, Config
@@ -73,55 +71,6 @@ class WSMsgType(str, Enum):
     FILE = "file"
 
 
-async def handler(ws: MelodieVisualizerProtocol, path):
-    global socks
-    if len(socks) >= MAX_ACTIVE_CONNECTIONS:
-        await ws.send(
-            json.dumps(
-                {
-                    "type": "error",
-                    "period": 0,
-                    "data": f"The number of connections exceeds the upper limit {MAX_ACTIVE_CONNECTIONS}. "
-                            f"Please shutdown unused webpage or modify the limit.",
-                    "modelState": UNCONFIGURED,
-                    "status": ERROR,
-                }
-            )
-        )
-        return
-    socks.add(ws)
-    visualize_result_queues[ws] = Queue(10)
-    while 1:
-        try:
-            content = await asyncio.wait_for(ws.recv(), timeout=0.05)
-            rec = json.loads(content)
-            cmd = rec["cmd"]
-            data = rec["data"]
-            if 0 <= cmd <= 10:
-                try:
-                    visualize_condition_queue_main.put((cmd, data, ws), timeout=1)
-                except:
-                    import traceback
-                    traceback.print_exc()
-            else:
-                raise NotImplementedError(cmd)
-        except (asyncio.TimeoutError, ConnectionRefusedError):
-            pass
-        except ConnectionClosedOK as e:
-            logger.info(e.reason)
-        if ws.closed:
-            socks.remove(ws)
-            visualize_result_queues.pop(ws)
-            logger.info(f"websocket connection {ws} is going offline...")
-            return
-        try:
-            while 1:
-                res = visualize_result_queues[ws].get(False)
-                await ws.send(res)
-        except queue.Empty:
-            pass
-
-
 class MelodieModelReset(BaseException):
     def __init__(self, ws: MelodieVisualizerProtocol = None):
         self.ws = ws
@@ -130,7 +79,6 @@ class MelodieModelReset(BaseException):
 def execute_only_enabled(func):
     """
     Execute the decorated function only if the Visualizer.enabled() is True.
-
     :param func:
     :return:
     """
@@ -155,13 +103,10 @@ class BaseVisualizer:
 
         self.params_dir = os.path.join(config.visualizer_tmpdir, 'params')
         self.sim_data_dir = os.path.join(config.visualizer_tmpdir, 'sim_data')
-        self.layout_data_dir = os.path.join(config.visualizer_tmpdir, "layout")
         if not os.path.exists(self.params_dir):
             os.makedirs(self.params_dir)
         if not os.path.exists(self.sim_data_dir):
             os.makedirs(self.sim_data_dir)
-        if not os.path.exists(self.layout_data_dir):
-            os.makedirs(self.layout_data_dir)
 
         self._model: "Model" = None
         self.params_manager: ParamsManager = ParamsManager()
@@ -176,6 +121,8 @@ class BaseVisualizer:
 
[…]
