### 2023-02-08 525dea8
Modified melodie visualizer dependencies
