### 2023-02-08 71b62bb
Modified melodie visualizer dependencies
