+++ Melodie/visualizer/__init__.py
@@ -1,6 +1,5 @@
-from .actions import Action
-from .params import FloatParam, ArrayParam, IntParam, Param, BoolParam, StringParam, SelectionParam
+from .actions import Action, Operation, DownloadOperation, ResponseConversionOperation, ShowChartWindowOperation
+from .params import FloatParam, ArrayParam, IntParam, Param, BoolParam, StringParam, SelectionParam, ParamsManager
 from .vis_charts import JSONBase
 from .vis_agent_series import AgentSeriesManager, AgentSeries
-from .visualizer import BaseVisualizer, NetworkVisualizer, Visualizer, MelodieModelReset
-from .ws_protocol import create_file_response, create_failed_response
+from .visualizer import BaseVisualizer, Visualizer, MelodieModelReset
