### 2023-02-15 ABM4ALL/Melodie#3 @JoelForamitti
I had the same issue (using Python 3.10.6) and installing sqlalchemy==1.4.41 as proposed above solved the issue. I would suggest to either fix this in the package requirements or ideally to support the latest version of sqlalchemy.
