### 2023-02-15 ABM4ALL/Melodie#7 @JoelForamitti
test_db fails
This issue is part of the following review: https://github.com/openjournals/joss-reviews/issues/5100. I followed the instructions here (https://abm4all.github.io/Melodie/html/installation.html) and also installed sqlalchemy as suggested here (https://github.com/ABM4ALL/Melodie/issues/6) and got the following failures and warnings after running pytest.

Here is my Melodie info response:

```
    Here is the information for Melodie
    ===================================
    Melodie Version  :  0.6.0
    Python Version   :  3.10.6
    Python Arch      :  ('64bit', 'ELF')
    Platform Detail  :  Linux-5.15.0-60-generic-x86_64-with-glibc2.35
    Cython Version   :  3.0.0a11
    Numpy Version    :  1.24.1

```

Here is the pytest response:

```
=================================================== FAILURES ====================================================
_______________________________________________ test_read_sqlite ________________________________________________

self = <sqlalchemy.engine.base.Connection object at 0x7fea22784550>
dialect = <sqlalchemy.dialects.sqlite.pysqlite.SQLiteDialect_pysqlite object at 0x7fea22786fe0>
constructor = <bound method DefaultExecutionContext._init_statement of <class 'sqlalchemy.dialects.sqlite.base.SQLiteExecutionContext'>>
statement = 'select * from simulator_scenarios', parameters = (), execution_options = immutabledict({})
args = ('select * from simulator_scenarios', []), kw = {}
branched = <sqlalchemy.engine.base.Connection object at 0x7fea22784550>
conn = <sqlalchemy.pool.base._ConnectionFairy object at 0x7fea22784190>
context = <sqlalchemy.dialects.sqlite.base.SQLiteExecutionContext object at 0x7fea22787580>
cursor = <sqlite3.Cursor object at 0x7fea22bb2b40>, evt_handled = False

    def _execute_context(
        self,
        dialect,
        constructor,
        statement,
        parameters,
        execution_options,
        *args,
        **kw
    ):
        """Create an :class:`.ExecutionContext` and execute, returning
        a :class:`_engine.CursorResult`."""

        branched = self
        if self.__branch_from:
            # if this is a "branched" connection, do everything in terms
            # of the "root" connection, *except* for .close(), which is
            # the only feature that branching provides
            self = self.__branch_from

        if execution_options:
            yp = execution_options.get("yield_per", None)
            if yp:
                execution_options = execution_options.union(
                    {"stream_results": True, "max_row_buffer": yp}
                )

        try:
            conn = self._dbapi_connection
            if conn is None:
                conn = self._revalidate_connection()

            context = constructor(
                dialect, self, conn, execution_options, *args, **kw
            )
        except (exc.PendingRollbackError, exc.ResourceClosedError):
            raise
        except BaseException as e:
            self._handle_dbapi_exception(
                e, util.text_type(statement), parameters, None, None
            )

        if (
            self._transaction
            and not self._transaction.is_active
            or (
                self._nested_transaction
                and not self._nested_transaction.is_active
            )
        ):
            self._invalid_transaction()

        elif self._trans_context_manager:
            TransactionalContext._trans_ctx_check(self)

        if self._is_future and self._transaction is None:
            self._autobegin()

        context.pre_exec()

        if dialect.use_setinputsizes:
            context._set_input_sizes()

        cursor, statement, parameters = (
            context.cursor,
            context.statement,
            context.parameters,
        )

        if not context.executemany:
            parameters = parameters[0]

        if self._has_events or self.engine._has_events:
            for fn in self.dispatch.before_cursor_execute:
                statement, parameters = fn(
                    self,
                    cursor,
                    statement,
                    parameters,
                    context,
                    context.executemany,
                )

        if self._echo:

            self._log_info(statement)

            stats = context._get_cache_stats()

            if not self.engine.hide_parameters:
                self._log_info(
                    "[%s] %r",
                    stats,
                    sql_util._repr_params(
                        parameters, batches=10, ismulti=context.executemany
                    ),
                )
            else:
                self._log_info(
                    "[%s] [SQL parameters hidden due to hide_parameters=True]"
                    % (stats,)
                )

        evt_handled = False
        try:
            if context.executemany:
                if self.dialect._has_events:
                    for fn in self.dialect.dispatch.do_executemany:
                        if fn(cursor, statement, parameters, context):
                            evt_handled = True
                            break
                if not evt_handled:
                    self.dialect.do_executemany(
                        cursor, statement, parameters, context
                    )
            elif not parameters and context.no_parameters:
                if self.dialect._has_events:
                    for fn in self.dialect.dispatch.do_execute_no_params:
                        if fn(cursor, statement, context):
                            evt_handled = True
                            break
                if not evt_handled:
                    self.dialect.do_execute_no_params(
                        cursor, statement, context
                    )
            else:
                if self.dialect._has_events:
                    for fn in self.dialect.dispatch.do_execute:
                        if fn(cursor, statement, parameters, context):
                            evt_handled = True
                            break
                if not evt_handled:
>                   self.dialect.do_execute(
                        cursor, statement, parameters, context
                    )

/home/joel/.virtualenvs/p3/lib/python3.10/site-packages/sqlalchemy/engine/base.py:1900:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

self = <sqlalchemy.dialects.sqlite.pysqlite.SQLiteDialect_pysqlite object at 0x7fea22786fe0>
cursor = <sqlite3.Cursor object at 0x7fea22bb2b40>, statement = 'select * from simulator_scenarios'
parameters = (), context = <sqlalchemy.dialects.sqlite.base.SQLiteExecutionContext object at 0x7fea22787580>

    def do_execute(self, cursor, statement, parameters, context=None):
>       cursor.execute(statement, parameters)
E       sqlite3.OperationalError: no such table: simulator_scenarios

/home/joel/.virtualenvs/p3/lib/python3.10/site-packages/sqlalchemy/engine/default.py:736: OperationalError

    def test_read_sqlite():
        path = os.path.join(resources_path, r"db\test.sqlite")
        conn_string = f"sqlite:///{path}"
>       resp = DatabaseService.execute_sql(conn_string, "select * from simulator_scenarios")

tests/infra/test_db.py:17:
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _
MelodieInfra/services/database.py:64: in execute_sql
    return pd.read_sql(sql, engine)
/home/joel/.virtualenvs/p3/lib/python3.10/site-packages/pandas/io/sql.py:590: in read_sql
    return pandas_sql.read_query(
/home/joel/.virtualenvs/p3/lib/python3.10/site-packages/pandas/io/sql.py:1560: in read_query
    result = self.execute(*args)
/home/joel/.virtualenvs/p3/lib/python3.10/site-packages/pandas/io/sql.py:1405: in execute
    return self.connectable.execution_options().execute(*args, **kwargs)
<string>:2: in execute
    ???
/home/joel/.virtualenvs/p3/lib/python3.10/site-packages/sqlalchemy/util/deprecations.py:402: in warned
    return fn(*args, **kwargs)
/home/joel/.virtualenvs/p3/lib/python3.10/site-packages/sqlalchemy/engine/base.py:3257: in execute
    return connection.execute(statement, *multiparams, **params)
/home/joel/.virtualenvs/p3/lib/python3.10/site-packages/sqlalchemy/engine/base.py:1365: in execute
    return self._exec_driver_sql(
/home/joel/.virtualenvs/p3/lib/python3.10/site-packages/sqlalchemy/engine/base.py:1669: in _exec_driver_sql
    ret = self._execute_context(
/home/joel/.virtualenvs/p3/lib/python3.10/site-packages/sqlalchemy/engine/base.py:1943: in _execute_context
    self._handle_dbapi_exception(
/home/joel/.virtualenvs/p3/lib/python3.10/site-packages/sqlalchemy/engine/base.py:2124: in _handle_dbapi_exception
    util.raise_(
/home/joel/.virtualenvs/p3/lib/python3.10/site-packages/sqlalchemy/util/compat.py:208: in raise_
    raise exception
/home/joel/.virtualenvs/p3/lib/python3.10/site-packages/sqlalchemy/engine/base.py:1900: in _execute_context
    self.dialect.do_execute(
_ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _ _

self = <sqlalchemy.dialects.sqlite.pysqlite.SQLiteDialect_pysqlite object at 0x7fea22786fe0>
cursor = <sqlite3.Cursor object at 0x7fea22bb2b40>, statement = 'select * from simulator_scenarios'
parameters = (), context = <sqlalchemy.dialects.sqlite.base.SQLiteExecutionContext object at 0x7fea22787580>

    def do_execute(self, cursor, statement, parameters, context=None):
>       cursor.execute(statement, parameters)
E       sqlalchemy.exc.OperationalError: (sqlite3.OperationalError) no such table: simulator_scenarios
E       [SQL: select * from simulator_scenarios]
E       (Background on this error at: https://sqlalche.me/e/14/e3q8)

/home/joel/.virtualenvs/p3/lib/python3.10/site-packages/sqlalchemy/engine/default.py:736: OperationalError
=============================================== warnings summary ================================================
tests/test_agent_manager.py:11
  /home/joel/2_Research/11 Reviews/2023_MELODIE/Melodie/tests/test_agent_manager.py:11: PytestCollectionWarning: cannot collect test class 'TestAgent' because it has a __init__ constructor (from: tests/test_agent_manager.py)
    class TestAgent(Agent):

tests/test_agent_manager.py:16
  /home/joel/2_Research/11 Reviews/2023_MELODIE/Melodie/tests/test_agent_manager.py:16: PytestCollectionWarning: cannot collect test class 'TestAgentToFilter' because it has a __init__ constructor (from: tests/test_agent_manager.py)
    class TestAgentToFilter(Agent):

Melodie/boost/agent_list.pyx:597
  /home/joel/2_Research/11 Reviews/2023_MELODIE/Melodie/Melodie/boost/agent_list.pyx:597: PytestCollectionWarning: cannot collect 'test_container' because it is not a function.
    cpdef test_container() except *:

tests/test_data_collector.py:29
  /home/joel/2_Research/11 Reviews/2023_MELODIE/Melodie/tests/test_data_collector.py:29: PytestCollectionWarning: cannot collect test class 'TestAgent' because it has a __init__ constructor (from: tests/test_data_collector.py)
    class TestAgent(Agent):

tests/test_data_collector.py:36
  /home/joel/2_Research/11 Reviews/2023_MELODIE/Melodie/tests/test_data_collector.py:36: PytestCollectionWarning: cannot collect test class 'TestEnv' because it has a __init__ constructor (from: tests/test_data_collector.py)
    class TestEnv(Environment):

tests/test_data_collector.py:41
  /home/joel/2_Research/11 Reviews/2023_MELODIE/Melodie/tests/test_data_collector.py:41: PytestCollectionWarning: cannot collect test class 'TestScenario' because it has a __init__ constructor (from: tests/test_data_collector.py)
    class TestScenario(Scenario):

tests/test_grid.py:15
  /home/joel/2_Research/11 Reviews/2023_MELODIE/Melodie/tests/test_grid.py:15: PytestCollectionWarning: cannot collect test class 'TestGridAgent' because it has a __init__ constructor (from: tests/test_grid.py)
    class TestGridAgent(GridAgent):

tests/test_model.py:9
  /home/joel/2_Research/11 Reviews/2023_MELODIE/Melodie/tests/test_model.py:9: PytestCollectionWarning: cannot collect test class 'TestAgent' because it has a __init__ constructor (from: tests/test_model.py)
    class TestAgent(Agent):

tests/test_model.py:16
  /home/joel/2_Research/11 Reviews/2023_MELODIE/Melodie/tests/test_model.py:16: PytestCollectionWarning: cannot collect test class 'TestModel' because it has a __init__ constructor (from: tests/test_model.py)
    class TestModel(Model):

tests/test_params.py:14
  /home/joel/2_Research/11 Reviews/2023_MELODIE/Melodie/tests/test_params.py:14: PytestCollectionWarning: cannot collect test class 'TestScenario' because it has a __init__ constructor (from: tests/test_params.py)
    class TestScenario(Scenario):

tests/test_table_generator.py:15
  /home/joel/2_Research/11 Reviews/2023_MELODIE/Melodie/tests/test_table_generator.py:15: PytestCollectionWarning: cannot collect test class 'TestModel' because it has a __init__ constructor (from: tests/test_table_generator.py)
    class TestModel(Model):

tests/test_table_generator.py:19
  /home/joel/2_Research/11 Reviews/2023_MELODIE/Melodie/tests/test_table_generator.py:19: PytestCollectionWarning: cannot collect test class 'TestDataframeLoader' because it has a __init__ constructor (from: tests/test_table_generator.py)
    class TestDataframeLoader(DataLoader):

tests/test_table_generator.py:27
  /home/joel/2_Research/11 Reviews/2023_MELODIE/Melodie/tests/test_table_generator.py:27: PytestCollectionWarning: cannot collect test class 'TestSimulator' because it has a __init__ constructor (from: tests/test_table_generator.py)
    class TestSimulator(Simulator):

tests/test_table_generator.py:35
  /home/joel/2_Research/11 Reviews/2023_MELODIE/Melodie/tests/test_table_generator.py:35: PytestCollectionWarning: cannot collect test class 'TestScenario' because it has a __init__ constructor (from: tests/test_table_generator.py)
    class TestScenario(Scenario):

tests/test_calibrator.py::test_calibrator
tests/test_data_collector.py::test_model_run
tests/test_new_trainer_algorithm.py::test_chrom_params_algorithm
tests/test_new_trainer_algorithm.py::test_chrom_params_algorithm
tests/test_table_generator.py::test_table_generator
  /home/joel/2_Research/11 Reviews/2023_MELODIE/Melodie/MelodieInfra/db/db.py:127: SADeprecationWarning: The Engine.table_names() method is deprecated and will be removed in a future release.  Please refer to Inspector.get_table_names(). (deprecated since: 1.4)
    logger.info(f"Database contains tables: {self.connection.table_names()}.")

tests/test_calibrator.py::test_calibrator
tests/test_data_collector.py::test_model_run
tests/test_new_trainer_algorithm.py::test_chrom_params_algorithm
tests/test_new_trainer_algorithm.py::test_chrom_params_algorithm
tests/test_table_generator.py::test_table_generator
  /home/joel/2_Research/11 Reviews/2023_MELODIE/Melodie/MelodieInfra/db/db.py:128: SADeprecationWarning: The Engine.table_names() method is deprecated and will be removed in a future release.  Please refer to Inspector.get_table_names(). (deprecated since: 1.4)
    table_names = list(self.connection.table_names())

tests/test_calibrator.py::test_calibrator
tests/test_new_trainer_algorithm.py::test_chrom_params_algorithm
  /home/joel/2_Research/11 Reviews/2023_MELODIE/Melodie/Melodie/utils/parallel_manager.py:86: DeprecationWarning: setDaemon() is deprecated, set the daemon attribute instead
    self.th_server.setDaemon(True)

tests/test_calibrator.py::test_calibrator
  /home/joel/2_Research/11 Reviews/2023_MELODIE/Melodie/Melodie/calibrator.py:329: RuntimeWarning: invalid value encountered in scalar divide
    cov = env_df[prop_name].std() / env_df[prop_name].mean()

tests/test_sqlite_performance.py::test_sqlite_performance
  /home/joel/.virtualenvs/p3/lib/python3.10/site-packages/_pytest/python.py:199: PytestReturnNotNoneWarning: Expected None, but tests/test_sqlite_performance.py::test_sqlite_performance returned '', which will be an error in a future version of pytest.  Did you mean to use `assert` instead of `return`?
    warnings.warn(

tests/test_sqlite_performance.py::test_select
  /home/joel/.virtualenvs/p3/lib/python3.10/site-packages/_pytest/python.py:199: PytestReturnNotNoneWarning: Expected None, but tests/test_sqlite_performance.py::test_select returned '', which will be an error in a future version of pytest.  Did you mean to use `assert` instead of `return`?
    warnings.warn(

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
============================================ short test summary info ============================================
FAILED tests/infra/test_db.py::test_read_sqlite - sqlalchemy.exc.OperationalError: (sqlite3.OperationalError) no such table: simulator_scenarios
=================================== 1 failed, 46 passed, 29 warnings in 5.96s ===================================
```
