+++ MelodieInfra/services/__init__.py
@@ -6,3 +6,4 @@
 
 from .files import *
 from .database import DatabaseService
+from .web import create_failed_response, create_json_response, create_file_response
