+++ MelodieInfra/parallel/parallel_worker.py
@@ -1,14 +1,11 @@
-#!/usr/bin/env python
-# -*- encoding: utf-8 -*-
+import argparse
 import base64
 import importlib
 import json
 import sys
 import time
-from typing import Dict, List, Tuple, Any, Type, Union, TYPE_CHECKING
-
 import rpyc
-import argparse
+from typing import Dict, Tuple, Any, Type, Union, TYPE_CHECKING
 
 import cloudpickle
 
