### 2023-02-15 ABM4ALL/Melodie#8 @JoelForamitti
Paper comments
This issue is part of the following review: https://github.com/openjournals/joss-reviews/issues/5100

Here are some comments on the paper:

- Statement of need
  - It would be helpful to not only describe how Melodie differs from Mesa and AgentPy, but also explain what need this satisfies. E.g. I understand that the environment component does in Melodie what is usually done in the Model component in Mesa and AgentPy, but what is the advantage of this design? The same applies to the second and third point.
  - To me it seems that a key novelty of this package is that is uses an sql database. Could you mention why you chose this approach and what the advantage is?
  - Please add a sentence on why you chose to work in Python, and mention that there are a large amount of ABM frameworks based on other programming languages.
- Overview
  - It is difficult to get an overview of how modelling in the package works by reading this section. I would suggest to show how an agent, a model, a secnario, and a visualizer is defined, the same way it is shown for the simulator.
  - Would it be possible to present a flowchart, showing how trainer, calibrator, scenario, etc. feed into each other?

Some typos:
- l12: mechanism -> mechanisms
- l15: ; (2) -> ; and (2)
- l57: data object -> data objects
