### 2023-02-15 #5100 @JoelForamitti
@SongminYu & @hzyrc6011, I have now also finished my review. My overall impression is that your package provides a lot of very interesting new functionalities, but also has a steeper learning curve than the other Python ABM frameworks. Thank you for your work, and please address the issues above.
