### 2023-02-15 ABM4ALL/Melodie#8 @JoelForamitti
Another novelty that you might want to highlight and explain the motivation is that this package uses cython components.
