### 2023-02-15 ABM4ALL/Melodie#7 @hzyrc6011
@JoelForamitti Thanks to your guidance about this.
I found that a backslash was used as file path splitter, so the path to the sqlite file was wrong on *nix systems.
I pushed a new version, hope that this time it will work.
