### 2023-02-16 ABM4ALL/Melodie#3 @rusu24edward
I tried again with Python 3.8.9. Here is my virtual environment:
```
astunparse        1.6.3
chardet           5.1.0
cloudpickle       2.2.1
contourpy         1.0.7
cycler            0.11.0
Cython            3.0.0a10
et-xmlfile        1.1.0
fonttools         4.38.0
greenlet          2.0.2
kiwisolver        1.4.4
matplotlib        3.6.3
Melodie           0.6.0
networkx          3.0
numpy             1.24.2
openpyxl          3.1.1
packaging         23.0
pandas            1.5.3
Pillow            9.4.0
pip               23.0
plumbum           1.8.1
pprintast         1.2.1
pyparsing         3.0.9
python-dateutil   2.8.2
pytz              2022.7.1
rpyc              5.3.0
scikit-opt        0.6.6
scipy             1.10.0
seaborn           0.12.2
setuptools        49.2.1
six               1.16.0
SQLAlchemy        2.0.3
typing_extensions 4.4.0
websockets        10.4
wheel             0.38.4
```

I get the same error.

I also did it by installing Melodie according to the developer installation docs. Same error.

---

Following your instructions to install sqlalchemy==1.4.41, I am able to run the tutorials using both python3.7 and python3.8 whether I install directly from pip or according to the developer's installation.
