### 2023-02-16 ABM4ALL/Melodie#11 @rusu24edward
run_analyzer not clear
In the covid contagion tutorial, I get the following output from `run_simulator.py`:
```
db.py:127 INFO Database contains tables: ['ID_AgeGroup', 'ID_HealthState', 'Parameter_AgeGroup_TransitionProb', 'Parameter_AgentParams', 'agents_result', 'environment_result', 'simulator_scenarios'].
db.py:131 INFO Database drops tables: ['ID_AgeGroup', 'ID_HealthState', 'Parameter_AgeGroup_TransitionProb', 'Parameter_AgentParams', 'agents_result', 'environment_result', 'simulator_scenarios'].
simulator.py:169 INFO Simulation started - id_scenario = 0, id_run = 0
simulator.py:221 INFO Simulation completed - id_scenario = 0, id_run = 0
    model-setup          0.041s
    model-run            0.09s
    data-collect         1.498s

simulator.py:169 INFO Simulation started - id_scenario = 1, id_run = 0
simulator.py:221 INFO Simulation completed - id_scenario = 1, id_run = 0
    model-setup          0.041s
    model-run            0.07s
    data-collect         1.55s

simulator.py:169 INFO Simulation started - id_scenario = 1, id_run = 1
simulator.py:221 INFO Simulation completed - id_scenario = 1, id_run = 1
    model-setup          0.04s
    model-run            0.074s
    data-collect         1.563s

simulator.py:257 INFO Simulator completed, time elapsed 5.145s.
```

I run `run_analyzer.py`, which generates files in the `/data/output` directory. It is not clear from the tutorial what is supposed to happen with `run_analyzer`, so please update for clarity.
