### 2023-02-17 ABM4ALL/Melodie#13 @rusu24edward
run analyzer in CovidContagionCalibrator tutorial
After `run_simulator.py`, I attempt `run_analyzer.py`, but it doesn't work. Here's the output:
```
Traceback (most recent call last):
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/sqlalchemy/engine/base.py", line 1900, in _execute_context
    self.dialect.do_execute(
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/sqlalchemy/engine/default.py", line 736, in do_execute
    cursor.execute(statement, parameters)
sqlite3.OperationalError: no such table: environment_calibrator_result_cov

Traceback (most recent call last):
  File "/Users/rusu1/tmp/Melodie/MelodieInfra/db/db.py", line 198, in read_dataframe
    return pd.read_sql(sql, self.connection)
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/pandas/io/sql.py", line 590, in read_sql
    return pandas_sql.read_query(
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/pandas/io/sql.py", line 1560, in read_query
    result = self.execute(*args)
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/pandas/io/sql.py", line 1405, in execute
    return self.connectable.execution_options().execute(*args, **kwargs)
  File "<string>", line 2, in execute
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/sqlalchemy/util/deprecations.py", line 402, in warned
    return fn(*args, **kwargs)
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/sqlalchemy/engine/base.py", line 3257, in execute
    return connection.execute(statement, *multiparams, **params)
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/sqlalchemy/engine/base.py", line 1365, in execute
    return self._exec_driver_sql(
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/sqlalchemy/engine/base.py", line 1669, in _exec_driver_sql
    ret = self._execute_context(
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/sqlalchemy/engine/base.py", line 1943, in _execute_context
    self._handle_dbapi_exception(
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/sqlalchemy/engine/base.py", line 2124, in _handle_dbapi_exception
    util.raise_(
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/sqlalchemy/util/compat.py", line 208, in raise_
    raise exception
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/sqlalchemy/engine/base.py", line 1900, in _execute_context
    self.dialect.do_execute(
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/sqlalchemy/engine/default.py", line 736, in do_execute
    cursor.execute(statement, parameters)
sqlalchemy.exc.OperationalError: (sqlite3.OperationalError) no such table: environment_calibrator_result_cov
[SQL: select * from environment_calibrator_result_cov]
(Background on this error at: https://sqlalche.me/e/14/e3q8)
Traceback (most recent call last):
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/sqlalchemy/engine/base.py", line 1900, in _execute_context
    self.dialect.do_execute(
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/sqlalchemy/engine/default.py", line 736, in do_execute
    cursor.execute(statement, parameters)
sqlite3.OperationalError: no such table: environment_calibrator_result_cov

Traceback (most recent call last):
  File "/Users/rusu1/tmp/Melodie/MelodieInfra/db/db.py", line 198, in read_dataframe
    return pd.read_sql(sql, self.connection)
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/pandas/io/sql.py", line 590, in read_sql
    return pandas_sql.read_query(
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/pandas/io/sql.py", line 1560, in read_query
    result = self.execute(*args)
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/pandas/io/sql.py", line 1405, in execute
    return self.connectable.execution_options().execute(*args, **kwargs)
  File "<string>", line 2, in execute
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/sqlalchemy/util/deprecations.py", line 402, in warned
    return fn(*args, **kwargs)
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/sqlalchemy/engine/base.py", line 3257, in execute
    return connection.execute(statement, *multiparams, **params)
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/sqlalchemy/engine/base.py", line 1365, in execute
    return self._exec_driver_sql(
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/sqlalchemy/engine/base.py", line 1669, in _exec_driver_sql
    ret = self._execute_context(
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/sqlalchemy/engine/base.py", line 1943, in _execute_context
    self._handle_dbapi_exception(
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/sqlalchemy/engine/base.py", line 2124, in _handle_dbapi_exception
    util.raise_(
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/sqlalchemy/util/compat.py", line 208, in raise_
    raise exception
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/sqlalchemy/engine/base.py", line 1900, in _execute_context
    self.dialect.do_execute(
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/sqlalchemy/engine/default.py", line 736, in do_execute
    cursor.execute(statement, parameters)
sqlalchemy.exc.OperationalError: (sqlite3.OperationalError) no such table: environment_calibrator_result_cov
[SQL: select * from environment_calibrator_result_cov]
(Background on this error at: https://sqlalche.me/e/14/e3q8)

Traceback (most recent call last):
  File "run_analyzer.py", line 9, in <module>
    analyzer.plot_calibration_process()
  File "/Users/rusu1/tmp/CovidContagionCalibrator/source/analyzer.py", line 51, in plot_calibration_process
    df = self.db.read_dataframe("environment_calibrator_result_cov")
  File "/Users/rusu1/tmp/Melodie/MelodieInfra/db/db.py", line 203, in read_dataframe
    raise MelodieExceptions.Data.AttemptingReadingFromUnexistedTable(table_name)
MelodieInfra.exceptions.exceptions.MelodieException: Table 'environment_calibrator_result_cov' does not in database. <Error ID 1503>
```

I get this error in both python3.7 and 3.8.
