### 2023-02-17 ABM4ALL/Melodie#14 @rusu24edward
Unclear how to access resulting tables from calibration tutorial
I ran `run_calibrator.py` in the CovidContagionCalibrator tutorial, and this was my output:
```
...
calibrator.py:350 INFO =======================Path 2 Generation 50/50=======================
parallel_worker.py:217 INFO Processor 7, chromosome 0, time: 0.193s
parallel_worker.py:217 INFO Processor 4, chromosome 5, time: 0.205s
parallel_worker.py:217 INFO Processor 5, chromosome 1, time: 0.238s
parallel_worker.py:217 INFO Processor 2, chromosome 2, time: 0.272s
parallel_worker.py:217 INFO Processor 1, chromosome 4, time: 0.261s
parallel_worker.py:217 INFO Processor 6, chromosome 3, time: 0.288s
parallel_worker.py:217 INFO Processor 0, chromosome 7, time: 0.277s
parallel_worker.py:217 INFO Processor 3, chromosome 6, time: 0.34s
parallel_worker.py:217 INFO Processor 7, chromosome 8, time: 0.294s
parallel_worker.py:217 INFO Processor 4, chromosome 9, time: 0.277s
parallel_worker.py:217 INFO Processor 5, chromosome 10, time: 0.303s
parallel_worker.py:217 INFO Processor 2, chromosome 11, time: 0.261s
parallel_worker.py:217 INFO Processor 1, chromosome 12, time: 0.306s
parallel_worker.py:217 INFO Processor 6, chromosome 13, time: 0.302s
parallel_worker.py:217 INFO Processor 0, chromosome 14, time: 0.312s
parallel_worker.py:217 INFO Processor 3, chromosome 15, time: 0.3s
parallel_worker.py:217 INFO Processor 7, chromosome 16, time: 0.302s
parallel_worker.py:217 INFO Processor 4, chromosome 17, time: 0.302s
parallel_worker.py:217 INFO Processor 5, chromosome 18, time: 0.302s
parallel_worker.py:217 INFO Processor 2, chromosome 19, time: 0.308s
parallel_worker.py:217 INFO Processor 1, chromosome 20, time: 0.305s
parallel_worker.py:217 INFO Processor 6, chromosome 21, time: 0.309s
parallel_worker.py:217 INFO Processor 0, chromosome 22, time: 0.303s
parallel_worker.py:217 INFO Processor 3, chromosome 23, time: 0.308s
parallel_worker.py:217 INFO Processor 7, chromosome 24, time: 0.295s
parallel_worker.py:217 INFO Processor 4, chromosome 25, time: 0.298s
parallel_worker.py:217 INFO Processor 5, chromosome 26, time: 0.294s
parallel_worker.py:217 INFO Processor 2, chromosome 27, time: 0.302s
parallel_worker.py:217 INFO Processor 1, chromosome 28, time: 0.288s
parallel_worker.py:217 INFO Processor 6, chromosome 29, time: 0.291s
parallel_worker.py:217 INFO Processor 0, chromosome 30, time: 0.29s
parallel_worker.py:217 INFO Processor 3, chromosome 31, time: 0.257s
parallel_worker.py:217 INFO Processor 7, chromosome 32, time: 0.288s
parallel_worker.py:217 INFO Processor 4, chromosome 33, time: 0.294s
parallel_worker.py:217 INFO Processor 5, chromosome 34, time: 0.294s
parallel_worker.py:217 INFO Processor 2, chromosome 35, time: 0.296s
parallel_worker.py:217 INFO Processor 1, chromosome 36, time: 0.295s
parallel_worker.py:217 INFO Processor 6, chromosome 37, time: 0.343s
parallel_worker.py:217 INFO Processor 0, chromosome 38, time: 0.338s
parallel_worker.py:217 INFO Processor 3, chromosome 39, time: 0.328s
parallel_worker.py:217 INFO Processor 7, chromosome 40, time: 0.329s
parallel_worker.py:217 INFO Processor 4, chromosome 41, time: 0.353s
parallel_worker.py:217 INFO Processor 5, chromosome 42, time: 0.336s
parallel_worker.py:217 INFO Processor 2, chromosome 43, time: 0.33s
parallel_worker.py:217 INFO Processor 1, chromosome 44, time: 0.277s
parallel_worker.py:217 INFO Processor 6, chromosome 45, time: 0.281s
parallel_worker.py:217 INFO Processor 0, chromosome 46, time: 0.283s
parallel_worker.py:217 INFO Processor 3, chromosome 47, time: 0.279s
parallel_worker.py:217 INFO Processor 7, chromosome 48, time: 0.228s
parallel_worker.py:217 INFO Processor 4, chromosome 49, time: 0.167s
calibrator.py:391 INFO ====================Time Elapsed: 2.0582101345062256s====================
parallel_manager.py:110 INFO terminated subworker <subprocess.Popen object at 0x7f87bc2e0f10>
parallel_manager.py:110 INFO terminated subworker <subprocess.Popen object at 0x7f87bc361a60>
parallel_manager.py:110 INFO terminated subworker <subprocess.Popen object at 0x7f87bc3585b0>
parallel_manager.py:110 INFO terminated subworker <subprocess.Popen object at 0x7f87bc487820>
parallel_manager.py:110 INFO terminated subworker <subprocess.Popen object at 0x7f87bc358190>
parallel_manager.py:110 INFO terminated subworker <subprocess.Popen object at 0x7f87bc459430>
parallel_manager.py:110 INFO terminated subworker <subprocess.Popen object at 0x7f87bc257880>
parallel_manager.py:110 INFO terminated subworker <subprocess.Popen object at 0x7f87bc2572e0>
server.py:123 INFO listener closed
server.py:207 INFO goodbye ('127.0.0.1', 51509)
server.py:207 INFO goodbye ('127.0.0.1', 51511)
server.py:271 INFO server has terminated
server.py:123 INFO listener closed
parallel_manager.py:113 INFO Server closed!
```

It's not clear if this was successful or not. I don't see the table or graph output anywhere.
