### 2023-02-17 ABM4ALL/Melodie#16 @rusu24edward
Covid Network Container Visual
I installed and launched `MelodieStudio`. Then I ran `run_simulator` from the CovidNetworkContagionVisual tutorial. The web page did not connect. Here is the output from the front-end and back-end, respectively:
```
 * Serving Flask app 'MelodieStudio.main'
 * Debug mode: off
WARNING: This is a development server. Do not use it in a production deployment. Use a production WSGI server instead.
 * Running on all addresses (0.0.0.0)
 * Running on http://127.0.0.1:8089
 * Running on http://128.15.165.38:8089
Press CTRL+C to quit
file /Users/rusu1/tmp/CovidNetworkContagionVisual/__pycache__/config.cpython-38.pyc.140460112340144 created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/__pycache__/config.cpython-38.pyc.140460112340144 moved, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/source/__pycache__/data_loader.cpython-38.pyc.140460098345696 created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/source/__pycache__/data_loader.cpython-38.pyc.140460098345696 moved, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/source/__pycache__/data_info.cpython-38.pyc.140460113164448 created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/source/__pycache__/data_info.cpython-38.pyc.140460113164448 moved, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/source/__pycache__/model.cpython-38.pyc.140460113165024 created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/source/__pycache__/model.cpython-38.pyc.140460113165024 moved, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/source/__pycache__/agent.cpython-38.pyc.140460113165312 created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/source/__pycache__/agent.cpython-38.pyc.140460113165312 moved, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/source/__pycache__/data_collector.cpython-38.pyc.140460113166176 created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/source/__pycache__/data_collector.cpython-38.pyc.140460113166176 moved, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/source/__pycache__/environment.cpython-38.pyc.140460113203392 created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/source/__pycache__/environment.cpython-38.pyc.140460113203392 moved, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/source/__pycache__/scenario.cpython-38.pyc.140460113203680 created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/source/__pycache__/scenario.cpython-38.pyc.140460113203680 moved, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/source/__pycache__/visualizer.cpython-38.pyc.140460113205120 created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/source/__pycache__/visualizer.cpython-38.pyc.140460113205120 moved, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/data/output/CovidNetworkContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/data/output/CovidNetworkContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/data/output/CovidNetworkContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/data/output/CovidNetworkContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/data/output/CovidNetworkContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/data/output/CovidNetworkContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/data/output/CovidNetworkContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/data/output/CovidNetworkContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/data/output/CovidNetworkContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/data/output/CovidNetworkContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/data/output/CovidNetworkContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/data/output/CovidNetworkContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/data/output/CovidNetworkContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/data/output/CovidNetworkContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/data/output/CovidNetworkContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/data/output/CovidNetworkContagionVisual.sqlite created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/data/output/CovidNetworkContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/data/output/CovidNetworkContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/data/output/CovidNetworkContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/data/output/CovidNetworkContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/data/output/CovidNetworkContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/data/output/CovidNetworkContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/data/output/CovidNetworkContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/data/output/CovidNetworkContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/data/output/CovidNetworkContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidNetworkContagionVisual/data/output/CovidNetworkContagionVisual.sqlite-journal created, now update!
127.0.0.1 - - [16/Feb/2023 18:35:49] "GET / HTTP/1.1" 301 -
127.0.0.1 - - [16/Feb/2023 18:35:49] "GET /index.html HTTP/1.1" 304 -
127.0.0.1 - - [16/Feb/2023 18:35:49] "GET /js/chunk-vendors.ec0d5ce6.js HTTP/1.1" 304 -
127.0.0.1 - - [16/Feb/2023 18:35:49] "GET /js/app.af103955.js HTTP/1.1" 304 -
127.0.0.1 - - [16/Feb/2023 18:35:50] "GET /api/tools/projectMeta HTTP/1.1" 200 -
127.0.0.1 - - [16/Feb/2023 18:35:50] "GET /js/703.11903911.js HTTP/1.1" 304 -
127.0.0.1 - - [16/Feb/2023 18:35:50] "GET /js/434.fd11c8bc.js HTTP/1.1" 304 -
127.0.0.1 - - [16/Feb/2023 18:35:50] "GET /js/305.4466cff8.js HTTP/1.1" 304 -
127.0.0.1 - - [16/Feb/2023 18:35:50] "GET /js/529.c47275d2.js HTTP/1.1" 304 -
127.0.0.1 - - [16/Feb/2023 18:35:50] "GET /api/charts/getLayout HTTP/1.1" 200 -
127.0.0.1 - - [16/Feb/2023 18:35:50] "GET /api/fs/getFile?fileName=README.md HTTP/1.1" 200 -
```

```
visualizer.py:130 INFO
====================================================================================================
Visualizer started at localhost:8765
====================================================================================================
 * Serving Flask app 'Melodie.visualizer.visualizer_server'
 * Debug mode: off
_internal.py:224 INFO WARNING: This is a development server. Do not use it in a production deployment. Use a production WSGI server instead.
 * Running on all addresses (0.0.0.0)
 * Running on http://127.0.0.1:8765
 * Running on http://128.15.165.38:8765
_internal.py:224 INFO Press CTRL+C to quit
db.py:127 INFO Database contains tables: ['ID_AgeGroup', 'ID_HealthState', 'Parameter_AgeGroup_TransitionProb', 'Parameter_AgentParams', 'simulator_scenarios'].
db.py:131 INFO Database drops tables: ['ID_AgeGroup', 'ID_HealthState', 'Parameter_AgeGroup_TransitionProb', 'Parameter_AgentParams', 'simulator_scenarios'].
simulator.py:275 INFO Simulator start up cost: 0.19620299339294434s
simulator.py:277 INFO Visualizer interactive paramerters for this scenario are: {}
db.py:127 INFO Database contains tables: [].
db.py:131 INFO Database drops tables: [].
simulator.py:294 INFO Scenario parameters: {'_parameters': [], 'manager': <Melodie.simulator.Simulator object at 0x7fbf63c65220>, 'id': 0, 'run_num': 1, 'period_num': 100, 'agent_num': 300, 'network_type': 'barabasi_albert_graph', 'network_param_k': 6, 'network_param_p': 0.2, 'network_param_m': 2, 'initial_infected_percentage': 0.1, 'young_percentage': 0.8, 'infection_prob': 0.1, 'transition_probs': {0: {'s1_s1': 0.9, 's1_s2': 0.099, 's1_s3': 0.0010000000000000009}, 1: {'s1_s1': 0.9, 's1_s2': 0.09, 's1_s3': 0.010000000000000009}}}
simulator.py:168 INFO Simulation started - id_scenario = 0, id_run = 0
vis_charts.py:287 INFO Chart manager reset!
visualizer.py:247 INFO Formatting current data takes:0.0897829532623291 seconds
visualizer.py:338 INFO in start
_internal.py:224 INFO 127.0.0.1 - - [16/Feb/2023 18:35:50] "GET / HTTP/1.1" 404 -
_internal.py:224 INFO 127.0.0.1 - - [16/Feb/2023 18:35:52] "GET / HTTP/1.1" 404 -
_internal.py:224 INFO 127.0.0.1 - - [16/Feb/2023 18:35:54] "GET / HTTP/1.1" 404 -
_internal.py:224 INFO 127.0.0.1 - - [16/Feb/2023 18:35:56] "GET / HTTP/1.1" 404 -
_internal.py:224 INFO 127.0.0.1 - - [16/Feb/2023 18:35:58] "GET / HTTP/1.1" 404 -
_internal.py:224 INFO 127.0.0.1 - - [16/Feb/2023 18:36:00] "GET / HTTP/1.1" 404 -
_internal.py:224 INFO 127.0.0.1 - - [16/Feb/2023 18:36:02] "GET / HTTP/1.1" 404 -
_internal.py:224 INFO 127.0.0.1 - - [16/Feb/2023 18:36:05] "GET / HTTP/1.1" 404 -
_internal.py:224 INFO 127.0.0.1 - - [16/Feb/2023 18:36:09] "GET / HTTP/1.1" 404 -
_internal.py:224 INFO 127.0.0.1 - - [16/Feb/2023 18:36:15] "GET / HTTP/1.1" 404 -
_internal.py:224 INFO 127.0.0.1 - - [16/Feb/2023 18:36:24] "GET / HTTP/1.1" 404 -
_internal.py:224 INFO 127.0.0.1 - - [16/Feb/2023 18:36:38] "GET / HTTP/1.1" 404 -
```
