### 2023-02-17 ABM4ALL/Melodie#8 @SongminYu
Hi @JoelForamitti, thanks for your valuable comments and suggestions. I have updated and pushed the manuscript to the JOSS-paper branch. The “Statement of Need” section is enhanced according to your suggestions. However, for the two overview points, I haven't successfully made the changes at the moment.

First, due to the length limit, we didn't include the details about how to set up each class. But this is introduced in detail in our documentation: (1) the tutorial section, and (2) the model gallery. We provided a link to the documentation and mentioned this in the manuscript.

Second, I also really wanted to add a figure to show how the modules are linked together. However, the figures I can think of are either too complicated and difficult to read or too simple and a bit misleading, especially when the trainer and the calibrator are also included. I tried to explain this in the tutorial in detail, as well as the documentation for each model in the gallery.
