### 2023-02-17 #5100 @SongminYu
Hi @JoelForamitti, thanks for your valuable comments. They are very helpful and we tried to revise the manuscript accordingly. I just replied to you in this [issue](https://github.com/ABM4ALL/Melodie/issues/8) and pushed a revised version to the repo. We are looking forward to your further comments and suggestions! Best regards, Songmin.
