+++ Melodie/visualizer/__init__.py
@@ -1,4 +1,4 @@
-from .actions import Action, Operation, DownloadOperation, ResponseConversionOperation, ShowChartWindowOperation
+from .actions import ToolbarAction, Operation, DownloadOperation, ResponseConversionOperation, ShowChartWindowOperation
 from .params import FloatParam, ArrayParam, IntParam, Param, BoolParam, StringParam, SelectionParam, ParamsManager
 from .vis_charts import JSONBase
 from .vis_agent_series import AgentSeriesManager, AgentSeries
