+++ requirements.txt
@@ -9,7 +9,7 @@ openpyxl
 xlrd==1.2.0
 networkx
 chardet~=4.0.0
-sqlalchemy
+sqlalchemy~=1.4
 astunparse
 pprintast
 sphinx
