+++ Melodie/visualizer/__init__.py
@@ -1,5 +1,3 @@
 from .actions import ToolbarAction, Operation, DownloadOperation, ResponseConversionOperation, ShowChartWindowOperation
-from .params import FloatParam, ArrayParam, IntParam, Param, BoolParam, StringParam, SelectionParam, ParamsManager
-from .vis_charts import JSONBase
 from .vis_agent_series import AgentSeriesManager, AgentSeries
 from .visualizer import BaseVisualizer, Visualizer, MelodieModelReset
