+++ MelodieInfra/models/__init__.py
@@ -3,6 +3,7 @@
 # @Author: Zhanyi Hou
 # @Email: 1295752786@qq.com
 # @File: __init__.py.py
+from .jsonbase import JSONBase
 from .table_files import *
 from .web import DataServiceStatus, DataServiceState
 from .database import DatabaseQueryRequest, DatabaseBasicRequest
