+++ MelodieInfra/lowcode/__init__.py
@@ -3,3 +3,5 @@
 # @Author: Zhanyi Hou
 # @Email: 1295752786@qq.com
 # @File: __init__.py.py
+
+from .params import ArrayParam, FloatParam, IntParam, StringParam, SelectionParam, ParamsManager, Param, BoolParam
