+++ tests/test_params.py
@@ -8,7 +8,7 @@ import json
 import pandas as pd
 
 from Melodie import Scenario
-from Melodie.visualizer.params import IntParam, ArrayParam, ParamsManager, FloatParam
+from MelodieInfra.lowcode.params import IntParam, ArrayParam, ParamsManager, FloatParam
 
 
 class TestScenario(Scenario):
