### 2023-02-23 8f332a0
modified installation.rst and fixed numpy data type error from issue #12
