### 2023-02-23 #5100 @SongminYu
Hi @jbytecode, we finished the revision comments by @JoelForamitti and received the confirmation ([issue #8](https://github.com/ABM4ALL/Melodie/issues/8)). We are now working on the other issues. We will come back to you once we finish them. Thanks!
