### 2023-02-28 a47b69f
modified the installation steps in the docs
