### 2023-03-02 ABM4ALL/Melodie#13 @rusu24edward
> our documentation will be improved later for clearity

I'm glad to hear you have it on your list. Clear instructions on running the software and working through the tutorials is an important aspect of the JOSS review.
