### 2023-03-02 ABM4ALL/Melodie#17 @rusu24edward
python versioning should be consistent
In some places you have updated the instructions to use python 3.8+, but others you have not. Here's what I was able to locate in the repo:

- [x] [Here](https://github.com/ABM4ALL/Melodie/blob/master/setup.py#L135) you should not have version of python less than 3.8.
- [x] [Here](https://github.com/ABM4ALL/Melodie/blob/master/setup.py#L165) should have version required >= 3.8
- [x] [Here](https://github.com/ABM4ALL/Melodie/blob/master/requirements.txt#L20) should show >= 3.8
- [x] I don't think you need [this line](https://github.com/ABM4ALL/Melodie/blob/master/build_requirements.txt#L1) anymore
- [x] I don't think you need [this line](https://github.com/ABM4ALL/Melodie/blob/master/.travis.yml#L4) anymore
