### 2023-03-04 ABM4ALL/Melodie#4 @SongminYu
Hi @rusu24edward, we have fixed this issue and also updated the installation document: Python >= 3.8 ([link](https://abm4all.github.io/Melodie/html/installation.html#user-installation)). Do you find other bugs? If not, you could also close this issue. Thanks!
