+++ setup.py
@@ -132,12 +132,10 @@ setuptools.setup(
         "Natural Language :: English",
         "Operating System :: Unix",
         "Operating System :: Microsoft :: Windows",
-        "Programming Language :: Python :: 3",
-        # 'Programming Language :: Python :: 3.6',
-        # 'Programming Language :: Python :: 3.7',
-        # 'Programming Language :: Python :: 3.8',
-        # 'Programming Language :: Python :: 3.9',
-        # 'Programming Language :: Python :: 3.10',
+        'Programming Language :: Python :: 3.8',
+        'Programming Language :: Python :: 3.9',
+        'Programming Language :: Python :: 3.10',
+        'Programming Language :: Python :: 3.11',
         "Topic :: Scientific/Engineering :: Artificial Intelligence",
     ],
     project_urls={
@@ -162,7 +160,7 @@ setuptools.setup(
         "flask-cors",
         "flask_sock"
     ],
-    python_requires=">=3.7",
+    python_requires=">=3.8",
     entry_points={"console_scripts": ["Melodie=Melodie.scripts.scripts:cli"]},
     include_package_data=True,
     ext_modules=ext_modules,
