### 2023-03-04 ABM4ALL/Melodie#14 @SongminYu
Hi @rusu24edward, thanks for the comment! We updated the document for this calibrator example for clarification. You can find it in the [Results](https://abm4all.github.io/Melodie/html/gallery/covid_contagion_calibrator.html#results) section. Please let us know if further revision is needed, or you could close this issue. Thanks!
