### 2023-03-04 ABM4ALL/Melodie#11 @SongminYu
Hi @rusu24edward, thanks for your comment! We added one more paragraph explaining the ``analyzer.py`` function. You can find the paragraph in the [DataCollection](https://abm4all.github.io/Melodie/html/tutorial.html#datacollector) section of the tutorial as below. Please let us know if further revision is needed, or you could close this issue. Thanks!

The added paragraph:

In the example project, we also prepared a simple ``analyzer.py`` file that produces two figures based on the results showing the population of different ``health_state``. The two figures will be saved in the ``data/output`` folder together with the ``.sqlite`` file. Similarly, users can also define other post-processing functions in the ``analyzer.py`` file.
