### 2023-03-04 #5100 @SongminYu
Hi @rusu24edward, thanks for your very detailed review and helpful comments! @hzyrc6011 and I managed to solve the following open issues today: [#3](https://github.com/ABM4ALL/Melodie/issues/3), [#4](https://github.com/ABM4ALL/Melodie/issues/4), [#11](https://github.com/ABM4ALL/Melodie/issues/11), [#13](https://github.com/ABM4ALL/Melodie/issues/13), [#14](https://github.com/ABM4ALL/Melodie/issues/14), [#17](https://github.com/ABM4ALL/Melodie/issues/17). Could you please look at them and see if you have further questions? If not, you could close them as complete. Thanks a lot! We still need a bit of time for the last issue [#16](https://github.com/ABM4ALL/Melodie/issues/16) of MelodieStudio.

Hi @JoelForamitti, thanks also for your comments. While fixing the issues above, we think the problems related to "Installation instructions" are solved. Could you have a look at the current status? If everything looks fine, could you check the last point and confirm the review is done?
