### 2023-03-11 ABM4ALL/Melodie#16 @hzyrc6011
Hi @rusu24edward , we have pushed a new version of `MelodieStudio` onto pypi, please re-install the `MelodieStudio` package, restart this demo, and try to visit `http://127.0.0.1:8089` by browser. Thanks!
