### 2023-03-12 #5100 @SongminYu
Hi @rusu24edward, we have fixed the last issue [16](https://github.com/ABM4ALL/Melodie/issues/16). Could you try and check the last "Functionality: Have the functional claims of the software been confirmed?" box if it works well? Then, we will release a new version of the package including all the revisions we made in this round. Thanks a lot!
