### 2023-03-13 ABM4ALL/Melodie#16 @rusu24edward
I've reinstalled my virtual environment from scratch using Python3.8 and installed `MelodieStudio` version 0.6.0. If I'm in the `Melodie` repo root, I can call `python3 -m MelodieStudio` and it returns
```
python3 -m MelodieStudio
 * Serving Flask app 'MelodieStudio.main'
 * Debug mode: off
WARNING: This is a development server. Do not use it in a production deployment. Use a production WSGI server instead.
 * Running on all addresses (0.0.0.0)
 * Running on http://127.0.0.1:8089
 * Running on http://128.15.167.49:8089
Press CTRL+C to quit
```
When I open `http://127.0.0.1:8089`, I just see a blank page. I think this is expected since this is the `Melodie` repo and not a tutorial, but please correct me if I'm wrong.

I hop over to CovidGridContagionVisual and run `python3 -m MelodieStudio`, and I receive the following error:
```
python3 -m MelodieStudio
Traceback (most recent call last):
  File "/Library/Frameworks/Python.framework/Versions/3.8/lib/python3.8/runpy.py", line 194, in _run_module_as_main
    return _run_code(code, main_globals, None,
  File "/Library/Frameworks/Python.framework/Versions/3.8/lib/python3.8/runpy.py", line 87, in _run_code
    exec(code, run_globals)
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/MelodieStudio/__main__.py", line 6, in <module>
    studio_main()
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/MelodieStudio/main.py", line 97, in studio_main
    init_config_manager(conf_folder)
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/MelodieStudio/utils/config_manager.py", line 153, in init_config_manager
    _config_manager = ConfigureManager(conf_folder)
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/MelodieStudio/utils/config_manager.py", line 126, in __init__
    self.basic_config: BasicConfig = BasicConfig(
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/MelodieStudio/utils/config_manager.py", line 35, in __init__
    self.load()
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/MelodieStudio/utils/config_manager.py", line 66, in load
    raise AttributeError(
AttributeError: Configure file /Users/rusu1/tmp/CovidGridContagionVisual/.melodie/studio/basic_config.json does not define 'CURRENT_VISUALIZER_HOST'
```

Same error when I try `CovidNetworkContagionVisual`.
