### 2023-03-14 ABM4ALL/Melodie#16 @rusu24edward
When I attempt to use `python3 -m MelodieStudio` in CovidGridContagionVisual, the browser is just blank, and nothing seems to happen when I run `run_simulator.py`. Here's the output:

```
CovidGridContagionVisual rusu1$ python3 -m MelodieStudio
 * Serving Flask app 'MelodieStudio.main'
 * Debug mode: off
WARNING: This is a development server. Do not use it in a production deployment. Use a production WSGI server instead.
 * Running on all addresses (0.0.0.0)
 * Running on http://127.0.0.1:8089
 * Running on http://128.15.170.86:8089
Press CTRL+C to quit
file /Users/rusu1/tmp/CovidGridContagionVisual/data/output/CovidGridContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidGridContagionVisual/data/output/CovidGridContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidGridContagionVisual/data/output/CovidGridContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidGridContagionVisual/data/output/CovidGridContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidGridContagionVisual/data/output/CovidGridContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidGridContagionVisual/data/output/CovidGridContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidGridContagionVisual/data/output/CovidGridContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidGridContagionVisual/data/output/CovidGridContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidGridContagionVisual/data/output/CovidGridContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidGridContagionVisual/data/output/CovidGridContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidGridContagionVisual/data/output/CovidGridContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidGridContagionVisual/data/output/CovidGridContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidGridContagionVisual/data/output/CovidGridContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidGridContagionVisual/data/output/CovidGridContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidGridContagionVisual/data/output/CovidGridContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidGridContagionVisual/data/output/CovidGridContagionVisual.sqlite created, now update!
file /Users/rusu1/tmp/CovidGridContagionVisual/data/output/CovidGridContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidGridContagionVisual/data/output/CovidGridContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidGridContagionVisual/data/output/CovidGridContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidGridContagionVisual/data/output/CovidGridContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidGridContagionVisual/data/output/CovidGridContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidGridContagionVisual/data/output/CovidGridContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidGridContagionVisual/data/output/CovidGridContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidGridContagionVisual/data/output/CovidGridContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidGridContagionVisual/data/output/CovidGridContagionVisual.sqlite-journal created, now update!
file /Users/rusu1/tmp/CovidGridContagionVisual/data/output/CovidGridContagionVisual.sqlite-journal created, now update!
```

```
CovidGridContagionVisual rusu1$ python3 run_simulator.py
visualizer.py:130 INFO
====================================================================================================
Visualizer started at localhost:8765
====================================================================================================
 * Serving Flask app 'Melodie.visualizer.visualizer_server'
 * Debug mode: off
_internal.py:224 INFO WARNING: This is a development server. Do not use it in a production deployment. Use a production WSGI server instead.
 * Running on all addresses (0.0.0.0)
 * Running on http://127.0.0.1:8765
 * Running on http://128.15.170.86:8765
_internal.py:224 INFO Press CTRL+C to quit
db.py:127 INFO Database contains tables: ['ID_AgeGroup', 'ID_HealthState', 'Parameter_AgeGroup_TransitionProb', 'Parameter_AgentParams', 'simulator_scenarios'].
db.py:131 INFO Database drops tables: ['ID_AgeGroup', 'ID_HealthState', 'Parameter_AgeGroup_TransitionProb', 'Parameter_AgentParams', 'simulator_scenarios'].
simulator.py:275 INFO Simulator start up cost: 0.3473050594329834s
simulator.py:277 INFO Visualizer interactive paramerters for this scenario are: {}
db.py:127 INFO Database contains tables: [].
db.py:131 INFO Database drops tables: [].
simulator.py:294 INFO Scenario parameters: {'_parameters': [], 'manager': <Melodie.simulator.Simulator object at 0x7fe0e0e03ee0>, 'id': 0, 'run_num': 1, 'period_num': 100, 'agent_num': 1000, 'grid_x_size': 100, 'grid_y_size': 100, 'initial_infected_percentage': 0.1, 'young_percentage': 0.8, 'infection_prob': 0.1, 'transition_probs': {0: {'s1_s1': 0.9, 's1_s2': 0.099, 's1_s3': 0.0010000000000000009}, 1: {'s1_s1': 0.9, 's1_s2': 0.09, 's1_s3': 0.010000000000000009}}}
simulator.py:168 INFO Simulation started - id_scenario = 0, id_run = 0
vis_charts.py:256 INFO Chart manager reset!
visualizer.py:247 INFO Formatting current data takes:0.0015358924865722656 seconds
visualizer.py:338 INFO in start
```
The same thing happens in the CovidNetworkContagionVisual.

Are you testing this yourself after you push a change? I'm just curious how it is working for you but not for me.
