### 2023-03-15 ABM4ALL/Melodie#16 @SongminYu
Hi @rusu24edward, this is weird... but the computer is never wrong... @hzyrc6011 tested again today and told me it was fine on his laptop. He is exploring some other reasons and will come back to you.
