### 2023-03-16 ABM4ALL/Melodie#16 @hzyrc6011
@jbytecode Thanks for your suggestion, and today I tried it on a ubuntu-22.04 server. The commands that I used for installing dependencies were:

```shell
# at the folder of /home/ubuntu

sudo add-apt-repository ppa:deadsnakes/ppa
sudo apt update
sudo apt install python3.8-* -y # installed the pre-built python package, and the building dependencies for python wheel c extensions

# clone Melodie to build the wheel
git clone https://github.com/ABM4ALL/Melodie.git

cd Melodie/
python3.8  -m pip install -r build_requirements.txt
python3.8 setup.py bdist_wheel
python3.8 -m pip install dist/Melodie-0.6.0-cp38-cp38-linux_x86_64.whl

# install MelodieStudio
python3.8 -m pip install MelodieStudio
```

And after starting the  simulator and studio, I visited `http://<address_of_server>:8089`, then the page was showed in the browser as the following picture:
![image](https://user-images.githubusercontent.com/36789453/225645226-39c69c71-0f33-4d95-a3c8-bb618aa17774.png)

If the problem still happens even after refreshing the visualization page, and the browser showed a completely blank page(even without any error code), could you please provide a screenshot inside the browser developer tool (Could be accessed by F12)? The screen shot for `Console` tab and `Network` tab will be helpful.

![Console page](https://user-images.githubusercontent.com/36789453/225646887-4f15f53a-c97f-4505-8e69-2f67d24879e8.png)

![Network page](https://user-images.githubusercontent.com/36789453/225647035-372d7e29-8e92-4b39-bbb3-6507c7444145.png)

Sorry again for this problem, we will struggle to find out how to fix it.
