### 2023-03-17 ABM4ALL/Melodie#16 @rusu24edward
Network tab:

Console tab:

Full output from console.
```
Loading failed for the <script> with source “https://cdn.bootcdn.net/ajax/libs/echarts/5.3.3/echarts.min.js”. [index.html:9:1](http://127.0.0.1:8089/index.html#/)
Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/container/src/aside.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/container/src/aside.mjs?
Source Map URL: aside.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/container/src/footer.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/container/src/footer.mjs?
Source Map URL: footer.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/container/src/header.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/container/src/header.mjs?
Source Map URL: header.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/container/src/main.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/container/src/main.mjs?
Source Map URL: main.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/container/style/css.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/container/style/css.mjs?
Source Map URL: css.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/main/style/css.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/main/style/css.mjs?
Source Map URL: css.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/header/style/css.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/header/style/css.mjs?
Source Map URL: css.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/aside/style/css.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/aside/style/css.mjs?
Source Map URL: css.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/menu/index.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/menu/index.mjs?
Source Map URL: index.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/menu/src/menu.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/menu/src/menu.mjs?
Source Map URL: menu.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/menu/src/utils/menu-bar.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/menu/src/utils/menu-bar.mjs?
Source Map URL: menu-bar.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/menu/src/utils/menu-item.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/menu/src/utils/menu-item.mjs?
Source Map URL: menu-item.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/menu/src/utils/submenu.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/menu/src/utils/submenu.mjs?
Source Map URL: submenu.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/utils/dom/aria.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/utils/dom/aria.mjs?
Source Map URL: aria.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/menu/src/menu-collapse-transition.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/menu/src/menu-collapse-transition.mjs?
Source Map URL: menu-collapse-transition.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/menu/src/sub-menu.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/menu/src/sub-menu.mjs?
Source Map URL: sub-menu.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/collapse-transition/index.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/collapse-transition/index.mjs?
Source Map URL: index.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/collapse-transition/src/collapse-transition.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/collapse-transition/src/collapse-transition.mjs?
Source Map URL: collapse-transition.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/menu/src/use-menu.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/menu/src/use-menu.mjs?
Source Map URL: use-menu.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/menu/src/use-menu-css-var.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/menu/src/use-menu-css-var.mjs?
Source Map URL: use-menu-css-var.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/menu/src/use-menu-color.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/menu/src/use-menu-color.mjs?
Source Map URL: use-menu-color.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/menu/src/menu-item2.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/menu/src/menu-item2.mjs?
Source Map URL: menu-item2.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/menu/src/menu-item.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/menu/src/menu-item.mjs?
Source Map URL: menu-item.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/menu/src/menu-item-group2.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/menu/src/menu-item-group2.mjs?
Source Map URL: menu-item-group2.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/menu/src/menu-item-group.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/menu/src/menu-item-group.mjs?
Source Map URL: menu-item-group.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/menu/style/css.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/menu/style/css.mjs?
Source Map URL: css.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/sub-menu/style/css.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/sub-menu/style/css.mjs?
Source Map URL: css.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/menu-item-group/style/css.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/menu-item-group/style/css.mjs?
Source Map URL: css.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/menu-item/style/css.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/menu-item/style/css.mjs?
Source Map URL: css.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/breadcrumb/index.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/breadcrumb/index.mjs?
Source Map URL: index.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/breadcrumb/src/breadcrumb2.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/breadcrumb/src/breadcrumb2.mjs?
Source Map URL: breadcrumb2.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/breadcrumb/src/breadcrumb.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/breadcrumb/src/breadcrumb.mjs?
Source Map URL: breadcrumb.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/tokens/breadcrumb.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/tokens/breadcrumb.mjs?
Source Map URL: breadcrumb.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/breadcrumb/src/breadcrumb-item2.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/breadcrumb/src/breadcrumb-item2.mjs?
Source Map URL: breadcrumb-item2.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/breadcrumb/src/breadcrumb-item.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/breadcrumb/src/breadcrumb-item.mjs?
Source Map URL: breadcrumb-item.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/breadcrumb/style/css.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/breadcrumb/style/css.mjs?
Source Map URL: css.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/breadcrumb-item/style/css.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/breadcrumb-item/style/css.mjs?
Source Map URL: css.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/mitt/dist/mitt.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/mitt/dist/mitt.mjs?
Source Map URL: mitt.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/tree/index.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/tree/index.mjs?
Source Map URL: index.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/tree/src/tree.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/tree/src/tree.mjs?
Source Map URL: tree.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/tree/src/model/tree-store.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/tree/src/model/tree-store.mjs?
Source Map URL: tree-store.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/tree/src/model/node.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/tree/src/model/node.mjs?
Source Map URL: node.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/tree/src/model/util.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/tree/src/model/util.mjs?
Source Map URL: util.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/tree/src/tree-node.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/tree/src/tree-node.mjs?
Source Map URL: tree-node.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/tree/src/tree-node-content.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/tree/src/tree-node-content.mjs?
Source Map URL: tree-node-content.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/tree/src/model/useNodeExpandEventBroadcast.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/tree/src/model/useNodeExpandEventBroadcast.mjs?
Source Map URL: useNodeExpandEventBroadcast.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/tree/src/model/useDragNode.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/tree/src/model/useDragNode.mjs?
Source Map URL: useDragNode.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/tree/src/model/useKeydown.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/tree/src/model/useKeydown.mjs?
Source Map URL: useKeydown.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/tree/style/css.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/tree/style/css.mjs?
Source Map URL: css.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/color-picker/index.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/color-picker/index.mjs?
Source Map URL: index.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/color-picker/src/index.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/color-picker/src/index.mjs?
Source Map URL: index.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/color-picker/src/components/alpha-slider.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/color-picker/src/components/alpha-slider.mjs?
Source Map URL: alpha-slider.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/color-picker/src/draggable.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/color-picker/src/draggable.mjs?
Source Map URL: draggable.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/utils/dom/position.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/utils/dom/position.mjs?
Source Map URL: position.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/color-picker/src/components/hue-slider.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/color-picker/src/components/hue-slider.mjs?
Source Map URL: hue-slider.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/color-picker/src/components/predefine.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/color-picker/src/components/predefine.mjs?
Source Map URL: predefine.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/color-picker/src/useOption.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/color-picker/src/useOption.mjs?
Source Map URL: useOption.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/color-picker/src/color.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/color-picker/src/color.mjs?
Source Map URL: color.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/color-picker/src/components/sv-panel.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/color-picker/src/components/sv-panel.mjs?
Source Map URL: sv-panel.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/color-picker/style/css.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/color-picker/style/css.mjs?
Source Map URL: css.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/select/index.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/select/index.mjs?
Source Map URL: index.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/select/src/select.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/select/src/select.mjs?
Source Map URL: select.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/tag/index.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/tag/index.mjs?
Source Map URL: index.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/tag/src/tag2.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/tag/src/tag2.mjs?
Source Map URL: tag2.mjs.map

Source map error: Error: unsupported protocol for sourcemap request webpack://MelodieStudio/node_modules/element-plus/es/components/tag/src/tag.mjs.map
Resource URL: webpack://MelodieStudio/./node_modules/element-plus/es/components/tag/src/tag.mjs?
Source Map URL: tag.mjs.map

[…]
