### 2023-03-20 ABM4ALL/Melodie#16 @hzyrc6011
@rusu24edward Thanks for the comment with problems, I found this was due to a misconfiguration of CDN in MelodieStudio,
and we tried to fixed this problem. Also, the fixes have already been tested on Windows, MacOS and Linux.

For Melodie, please install the latest commit on master branch;
For MelodieStudio, please install this wheel: https://github.com/ABM4ALL/MelodieStudio/releases/tag/0.7.0.dev1

After starting the Visualizer and MelodieStudio, the MelodieStudio will show the following message. Note that if melodieStudio started before the visualizer, it may raise `ConnectionRefusedError` because it cannot connect to the visualizer.
```bash
Melodie Studio is running on:
- Gateway:          8089
- Studio Service:   52422
- Visualizer:       127.0.0.1:8765

Please visit this url http://localhost:8089 to open the visualizer.

If MelodieStudio cannot connect to the visualizer, please check if the visualizer host in webpage is corresponding to the visualizer configuration.

 * Serving Flask app 'MelodieStudio.main'
 * Debug mode: off
```

Following this message, please visit `http://localhost:8089` page to see if this bug was fixed. Thank you again for your patience in reporting this problem!
