### 2023-03-20 ABM4ALL/Melodie#16 @rusu24edward
I followed these instructions, and I was able to get the Network Visual tutorial to work. Thanks for your hard work to fix this! Very cool visualization, definitely worth the effort.

I did try the Covid Grid Visual tutorial, and I got the following error:
```
python3 -m MelodieStudio

Melodie Studio is running on:
- Gateway:          8089
- Studio Service:   53713
- Visualizer:       127.0.0.1:8765

Please visit this url http://localhost:8089 to open the visualizer.

If MelodieStudio cannot connect to the visualizer, please check if the visualizer host in webpage is corresponding to the visualizer configuration.

 * Serving Flask app 'MelodieStudio.main'
 * Debug mode: off
Traceback (most recent call last):
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/MelodieStudio/gateway/gateway.py", line 550, in _send_message
    self.connection.send(bytes([0x80 + opcode]))
BrokenPipeError: [Errno 32] Broken pipe
Traceback (most recent call last):
  File "/Users/rusu1/tmp/Melodie/v_melodie_38/lib/python3.8/site-packages/MelodieStudio/gateway/gateway.py", line 550, in _send_message
    self.connection.send(bytes([0x80 + opcode]))
OSError: [Errno 9] Bad file descriptor
```
