### 2023-03-20 ABM4ALL/Melodie#16 @SongminYu
@rusu24edward Thanks a lot! We will have a look at this issue. We also noticed that there are still some aspects of MelodieStudio that need improvements. We will also try to fix them and release a new version!
