### 2023-03-21 #5100 @SongminYu
@jbytecode Thanks a lot for your effort! I want to ask about one more thing: in JOSS, is it possible to mark both me and @hzyrc6011 as the corresponding authors? My intention was like that, with twice "corresponding: true". But in the generated pdf, it is not clear.
