+++ paper/paper.md
@@ -10,7 +10,6 @@ authors:
   - name: Songmin Yu
     orcid: 0000-0001-6062-4382
     equal-contrib: true 
-    corresponding: true
     affiliation: 1
   - name: Zhanyi Hou
     orcid: 0000-0001-8524-5370
