### 2023-03-21 59f13b0
Joss paper (#20)
* update bibliography

* update paper

* update paper

* update paper

* update paper
