+++ paper/paper.bib
@@ -1,7 +1,7 @@
 
 @article{AgentPy,
     author    = {Joël Foramitti},
-    title     = {AgentPy: A package for agent-based modeling in Python},
+    title     = {AgentPy: A package for agent-based modeling in {P}ython},
     year      = {2021},
     journal   = {Journal of Open Source Software},
     publisher = {The Open Journal},
@@ -33,13 +33,13 @@
 
 @InProceedings{Mesa,
     author    = {Kazil, Jackie and Masad, David and Crooks, Andrew"},
-    title     = {Utilizing Python for Agent-Based Modeling: The Mesa Framework},
+    title     = {Utilizing {P}ython for Agent-Based Modeling: The {M}esa Framework},
     year      = {2020},
     editor    = {Thomson, Robert and Bisgin, Halil and Dancy, Christopher and Hyder, Ayaz and Hussain, Muhammad},
     booktitle = {Social, Cultural, and Behavioral Modeling},
     publisher = {Springer International Publishing},
     pages     = {308--317},
-    abstract  = {Mesa is an agent-based modeling framework written in Python. Originally started in 2013, it was created to be the go-to tool in for researchers wishing to build agent-based models with Python. Within this paper we present Mesa's design goals, along with its underlying architecture. This includes its core components: 1) the model (Model, Agent, Schedule, and Space), 2) analysis (Data Collector and Batch Runner) and the visualization (Visualization Server and Visualization Browser Page). We then discuss how agent-based models can be created in Mesa. This is followed by a discussion of applications and extensions by other researchers to demonstrate how Mesa design is decoupled and extensible and thus creating the opportunity for a larger decentralized ecosystem of packages that people can share and reuse for their own needs. Finally, the paper concludes with a summary and discussion of future development areas for Mesa.},
+    abstract  = {Mesa is an agent-based modeling framework written in {P}ython. Originally started in 2013, it was created to be the go-to tool in for researchers wishing to build agent-based models with Python. Within this paper we present Mesa's design goals, along with its underlying architecture. This includes its core components: 1) the model (Model, Agent, Schedule, and Space), 2) analysis (Data Collector and Batch Runner) and the visualization (Visualization Server and Visualization Browser Page). We then discuss how agent-based models can be created in Mesa. This is followed by a discussion of applications and extensions by other researchers to demonstrate how Mesa design is decoupled and extensible and thus creating the opportunity for a larger decentralized ecosystem of packages that people can share and reuse for their own needs. Finally, the paper concludes with a summary and discussion of future development areas for Mesa.},
     address   = {Cham},
     isbn      = {978-3-030-61255-9}
 }
