### 2023-03-21 #5100 @jbytecode
@openjournals/dev  - We encountered a problem with pdf rendering. The pdf above renders `[@agentpy]` to `[Foramitti, 2021]` whereas it is correctly rendered as `(Foramitti, 2021)` in `whedon.theoj.org` site. What to do here?
