### 2023-03-22 #5100 @jbytecode
@SongminYu - While we seek for a solution to the rendering problem, you can perform the tasks listed in [Additional Author Tasks After Review is Complete](https://github.com/openjournals/joss-reviews/issues/5100#issuecomment-1477352659).  Thank you in advance.
