### 2023-03-22 #5100 @xuanxu
The cause of that pdf rendering is some ambiguity in the Markdown syntax. The link info is being duplicated by adding a Markdown link **and** a footnote with the link. Removing the footnote fixes the issue.

Another workaround is to use syntax like `[[@Mesa]]{}[^4]`, which removes the ambiguity.
I'd suggest removing the footnotes. They don't really add any value.
