+++ paper/paper.md
@@ -41,10 +41,10 @@ ABMs are also flexible to consider agents’
 (2) bounded rationality and adaptation behavior based on psychological and behavioral studies.
 
 `Melodie` is a general framework for developing agent-based models (ABMs) in Python. 
-It is published and maintained on the GitHub organization page of [`ABM4ALL`](https://github.com/ABM4ALL)[^1],
+It is published and maintained on the GitHub organization page of [`ABM4ALL`](https://github.com/ABM4ALL),
 a developing community among agent-based modelers for sharing ideas and resources.
-Together with the code [repository](https://github.com/ABM4ALL/Melodie)[^2], 
-we also published the [documentation](https://abm4all.github.io/Melodie/html/index.html)[^3] of `Melodie`, 
+Together with the code [repository](https://github.com/ABM4ALL/Melodie), 
+we also published the [documentation](https://abm4all.github.io/Melodie/html/index.html) of `Melodie`, 
 including a tutorial explaining how a minimum example - an agent-based covid contagion model - 
 can be developed with `Melodie` step by step.
 
@@ -179,9 +179,7 @@ a comparison between `Melodie` and the two packages is provided in the documenta
 based on the same covid contagion model developed with all the three packages.
 
 
-[^1]: Link to `ABM4ALL`: [https://github.com/ABM4ALL](https://github.com/ABM4ALL).
-[^2]: Link to `Melodie` repository: [https://github.com/ABM4ALL/Melodie](https://github.com/ABM4ALL/Melodie).
-[^3]: Link to `Melodie` documentation: [https://abm4all.github.io/Melodie/html/index.html](https://abm4all.github.io/Melodie/html/index.html).
+
 
 
 # Acknowledgements
