### 2023-03-25 ABM4ALL/Melodie#16 @SongminYu
Hi @rusu24edward, we just released a new version 0.7.0 of Melodie and MelodieStudio. Now both examples of Network and Grid can be visualized! Thanks a lot!
