### 2023-03-25 #5100 @SongminYu
Hi @jbytecode, we have checked the boxes. The new version of Melodie 0.7.0 is released and also uploaded to Zenodo (10.5281/zenodo.7770202). Please let me know if there is anything else that I should do :) Thanks a lot for the support from all the beginning :)
