### 2023-03-26 #5100 @jbytecode
@SongminYu - Could you please change the line 18 of bibliography:

`title   = {An Agent-Based Framework for Policy Simulation: Modeling Heterogeneous Behaviors With Modified Sigmoid Function and Evolutionary Training},`

to

`title   = {An Agent-Based Framework for Policy Simulation: Modeling Heterogeneous Behaviors With Modified {S}igmoid Function and Evolutionary Training},`

to capitalize the name of the Sigmoid function?

Thank you in advance.
