### 2023-03-26 f00a699
Joss paper (#23)
* update bibliography

* update paper

* update paper

* update paper

* update paper

* update paper

* update paper

* update paper

* update paper

* update paper

* update bibtex
