+++ paper/paper.bib
@@ -15,7 +15,7 @@
 @article{Yu,
   author  = {Yu, Songmin},
   journal = {IEEE Transactions on Computational Social Systems},
-  title   = {An Agent-Based Framework for Policy Simulation: Modeling Heterogeneous Behaviors With Modified Sigmoid Function and Evolutionary Training},
+  title   = {An Agent-Based Framework for Policy Simulation: Modeling Heterogeneous Behaviors With Modified {S}igmoid Function and Evolutionary Training},
   year    = {2022},
   volume  = {},
   number  = {},
