+++ paper/paper.md
@@ -34,7 +34,7 @@ The interactions can be water molecules forming a vortex,
 ants searching for food, or people trading stocks in the market.
 
 Agents’ interactions can bring emergent properties to a system and turn it into a complex system. 
-To model such mechanisms is usually the core reason for using ABMs. 
+The core reason for using ABMs is usually to model such mechanisms. 
 Besides, taking social economic systems as example, 
 ABMs are also flexible to consider agents’ 
 (1) heterogeneity (e.g., wealth, risk attitude, preference, decision-making rule, etc.) based on micro-data; and
@@ -44,7 +44,7 @@ ABMs are also flexible to consider agents’
 It is published and maintained on the GitHub organization page of [`ABM4ALL`](https://github.com/ABM4ALL),
 a developing community among agent-based modelers for sharing ideas and resources.
 Together with the code [repository](https://github.com/ABM4ALL/Melodie), 
-we also published the [documentation](https://abm4all.github.io/Melodie/html/index.html) of `Melodie`, 
+we have also published the [documentation](https://abm4all.github.io/Melodie/html/index.html) of `Melodie`, 
 including a tutorial explaining how a minimum example - an agent-based covid contagion model - 
 can be developed with `Melodie` step by step.
 
@@ -59,42 +59,42 @@ Modelers can also benefit from the wealth of packages available for statistical 
 Following the tradition of [`NetLogo`](https://ccl.northwestern.edu/netlogo/) [@Netlogo],
 `Mesa` and `AgentPy` both support interactive simulation but with different focus and style. 
 
-In summary, `Melodie` distinguishes from `Mesa` and `AgentPy` in the following aspects.
+In summary, `Melodie` is distinguished from `Mesa` and `AgentPy` by the following aspects.
 
 First, `Melodie` separates an `environment` component from the `model` in `Mesa` and `AgentPy` for two dedicated tasks: 
 (1) storing the macro-level variables; and (2) coordinating the agents' decision-making and interaction processes. 
-With a separated `environment` component, the "storyline" of the model can be summarized under a `run` function in the `model` clearly. 
-Compared to the use of `scheduler` and `step` functions in different layers in `Mesa` and bundling the behavior functions of agents to the `AgentList` in `AgentPy`, 
-we think this is easier for the users to understand the logic. 
+With a separated `environment` component, the "storyline" of the model can be clearly summarized under a `run` function in the `model`. 
+Compared to the use of `scheduler` and `step` functions in different layers in `Mesa` and the bundling of the behavior functions of agents to the `AgentList` in `AgentPy`, 
+we think this makes it easier for users to understand the logic. 
 
 Second, `Melodie` enhances the `data_collector` component with higher configurability.
-The users can define functions for parsing specific data structure from the `agents` and the `environment`.
+Users can define functions for parsing specific data structure from the `agents` and the `environment`.
 For example, in a financial ABM, the transactions could be saved in the `environment` as `List[Transaction]`. 
-Then, in the `data_collector`, the users can define a function `collect_transaction_data()` to first parse the list 
-and then save the results into the database.
+Then, in the `data_collector`, the users could define a function `collect_transaction_data()` to first parse the list 
+and to then save the results into the database.
 
-Third, `Melodie` has a wider infrastructure coverage and provides dedicated modules for scenario management.
+Third, `Melodie` has wider infrastructure coverage and provides dedicated modules for scenario management.
 
-* All the input data are first registered and then loaded by a `data_loader` object into a `scenario` object. Then, as the input data container, `scenario` can be accessed by the `model` and its components, including `environment`, `data_collector`, and each `agent`.
[…]
