### 2023-03-26 #5100 @danielskatz
@SongminYu - this paper has been published correctly (though there was an error in the JOSS scripting that caused it to look like it hadn't) - see https://doi.org/10.21105/joss.05100
